landgpt/requirements.txt
# Core runtime
numpy>=1.26
torch>=2.3
lightning>=2.3
hydra-core==1.3.2
hydra-colorlog>=1.2
pandas>=2.2
xarray>=2024.1
netCDF4>=1.6
scipy>=1.11
rasterio>=1.3
albumentations>=1.4
torchvision>=0.19
einops>=0.8
pyrootutils>=1.0
matplotlib>=3.9
rich>=13.7
tensorboard>=2.17
fire>=0.6
pysteps>=1.10
cartopy>=0.23
torchinfo>=1.8
jupyter>=1.0
ipykernel>=6.29
dask[complete]>=2024.5
PyYAML>=6.0
